requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
Faker==8.1.2
Pillow==8.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
